requirements.txt
diffusers @ git+https://github.com/huggingface/diffusers.git@v0.24.0
transformers @ git+https://github.com/huggingface/transformers.git@v4.36.0
accelerate @ git+https://github.com/huggingface/accelerate.git@v0.25.0
numpy
torchsde
torchvision
ninja
safetensors
Pillow
xformers
triton
bitsandbytes
omegaconf
fastapi
pydantic
uvicorn
aiohttp
gputil